# Bump up Gradle to v8.8 for better build performance