--- a/buildSrc/build.gradle.kts
+++ b/buildSrc/build.gradle.kts
@@ -5,5 +5,5 @@ repositories {
 
 
 dependencies {
-    testImplementation 'junit:junit:[4.12,)'
+    testImplementation("junit:junit:[4.12,)")
 }