--- a/buildSrc/build.gradle.kts
+++ b/buildSrc/build.gradle.kts
@@ -1,9 +1,14 @@
+plugins {
+    `kotlin-dsl`
+}
+
 repositories {
     mavenLocal()
     mavenCentral()
 }
 
-
 dependencies {
+    implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:1.8.0")
+    implementation("com.diffplug.spotless:spotless-plugin-gradle:6.16.0")
     testImplementation("junit:junit:[4.12,)")
 }