--- a/settings.gradle.kts
+++ b/settings.gradle.kts
@@ -1,0 +1,7 @@
+rootProject.name = "elasticsearch-sudachi"
+include(
+    "spi",
+    "testlib",
+    "subplugin",
+    "integration",
+)