--- a/settings.gradle.kts
+++ b/settings.gradle.kts
@@ -1,4 +1,5 @@
 rootProject.name = "elasticsearch-sudachi"
+
 include(
     "spi",
     "testlib",