… .github/workflows/build.yml
… build.gradle
… gradle/wrapper/gradle-wrapper.jar
… integration/build.gradle
… spi/build.gradle
… spi/build.gradle.kts
… src/test/java/com/worksap/nlp/elasticsearch/sudachi/index/SearchEngineEnv.kt
… subplugin/build.gradle
… subplugin/build.gradle.kts
… testlib/build.gradle
… testlib/build.gradle.kts