## benchmark result computed by hyperfine (n=10)

Here is the comparison of `hyperfine --warmup 1 './gradlew clean build --no-build-cache'`
Actually the biggest bottleneck is `downloadIcuPlugin` task that downloads a huge file, so this PR does not contribute on performance improvements so much.

### `develop` (8.1.1)

`Time (mean ± σ):     27.141 s ±  1.430 s`

### this PR (8.8)

`Time (mean ± σ):     24.978 s ±  2.194 s`